my-mcp-server/requirements.txt
mcp[cli]
httpx
anthropic
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
html2text
openai
requests
pandas
google-genai
google-generativeai
beautifulsoup4
